+++ src/app/Balloon.App.Elasticsearch/Console/Elasticsearch.php
@@ -49,14 +49,22 @@ class Elasticsearch
     protected $fs;
 
     /**
+     * Bulk.
+     *
+     * @var int
+     */
+    protected $bulk = 200;
+
+    /**
      * Constructor.
      */
-    public function __construct(GetOpt $getopt, Server $server, Scheduler $scheduler, LoggerInterface $logger)
+    public function __construct(GetOpt $getopt, Server $server, Scheduler $scheduler, LoggerInterface $logger, array $config = [])
     {
         $this->logger = $logger;
         $this->getopt = $getopt;
         $this->fs = $server->getFilesystem();
         $this->scheduler = $scheduler;
+        $this->setOptions($config);
     }
 
     /**
@@ -70,18 +78,59 @@ class Elasticsearch
             return false;
         }
 
-        $this->logger->info('reindex elasticsearch requested, this is an async task, be sure you have enaugh workers running.', [
+        $total = $this->fs->countNodes();
+        $this->logger->info('reindex elasticsearch, nodes left: {total}/{total}', [
             'category' => get_class($this),
+            'total' => $total,
         ]);
 
+        $stack = [];
+        $done = 0;
+
         foreach ($this->fs->findNodesByFilter([]) as $node) {
-            $this->scheduler->addJob(Job::class, [
+            $stack[] = $this->scheduler->addJob(Job::class, [
                 'id' => $node->getId(),
                 'action' => Job::ACTION_CREATE,
             ]);
+
+            if (count($stack) >= $this->bulk) {
+                $this->logger->info('waiting for ['.$this->bulk.'] jobs to be finished', [
+                    'category' => get_class($this),
+                ]);
+
+                $done += $this->bulk;
+                $this->logger->info('reindex elasticsearch {percent}, nodes left: {done}/{total}', [
+                    'category' => get_class($this),
+                    'percent' => round($done / $total * 100, 1).'%',
+                    'total' => $total,
+                    'done' => $done,
+                ]);
+
+                $this->scheduler->waitFor($stack);
+                $stack = [];
+            }
         }
 
         return true;
+    }
+
+    /**
+     * Set options.
+     */
+    public function setOptions(array $config = [])
+    {
+        foreach ($config as $key => $value) {
+            switch ($key) {
+                case 'bulk':
+                    $this->{$key} = (int) $value;
+
+                break;
+                default:
+                    throw new InvalidArgumentException('invalid option '.$key.' given');
+            }
+        }
+
+        return $this;
     }
 
     /**